+++ Engine/Components/Camera/ICameraTarget.cs
@@ -13,7 +13,7 @@ public interface ICameraTarget
     /// </summary>
     public VectorInt Size { get; }
 
-    internal FrameBuffer Buffer { get; }
+    internal IRenderTarget RenderTarget { get; }
 
     internal void Update();
 }